# sdk/go/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
servers:
- url: /
paths:
...
  /{app_id}/oauth2/client-metadata.json:
    get:
      description: Get an Application client metadata.
      operationId: app_metadata
      parameters:
      - explode: false
        in: path
        name: app_id
        required: true
        schema:
          title: App Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppMetadataResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: App Metadata
      tags:
      - Authorization Server
  /{app_id}/oauth2/token:
    post:
      description: Generate a new token based on the request parameters.
      operationId: token
      parameters:
      - explode: false
        in: path
        name: app_id
        required: true
        schema:
          title: App Id
          type: string
        style: simple
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/TokenRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Token
      tags:
      - Authorization Server
  /{app_id}/oauth2/token_exchange:
    post:
      description: Do a token exchange for an app.
      operationId: token_exchange
      parameters:
      - explode: false
        in: path
        name: app_id
        required: true
        schema:
          title: App Id
          type: string
        style: simple
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/TokenExchangeRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Token Exchange
      tags:
      - Authorization Server
  /oauth2/introspect:
    post:
      description: Introspect a token and evaluate it against the requested tools.
      operationId: introspect
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Body_introspect"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenIntrospectResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Introspect
      tags:
      - Authorization Server
...
  /trace/llm/call_start:
    post:
      description: Record the start of an LLM call for the authenticated agent.
      operationId: trace_llm_call_start
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LLMCallStartedRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LLMCallStartedEvent"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Trace Llm Call Start
      tags:
      - Traces
  /trace/llm/call_end:
    post:
      description: Record the end of an LLM call for the authenticated agent.
      operationId: trace_llm_call_end
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LLMCallEndedRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LLMCallEndedEvent"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Trace Llm Call End
      tags:
      - Traces
...
  /k8s/namespaces/{namespace}/get_mas_by_app_host:
    get:
      operationId: get_k8s_mas_by_app_host
      parameters:
      - explode: false
        in: path
        name: namespace
        required: true
        schema:
          title: Namespace
          type: string
        style: simple
      - explode: true
        in: query
        name: app_host
        required: true
        schema:
          title: App Host
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/K8sMultiAgentSystemCRDViewModel"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get K8S Mas By App Host
      tags:
      - Kubernetes Resources
  /k8s/namespaces/{namespace}/get_mas_by_app_workload:
    get:
      operationId: get_k8s_mas_by_app_workload
      parameters:
      - explode: false
        in: path
        name: namespace
        required: true
        schema:
          title: Namespace
          type: string
        style: simple
      - explode: true
        in: query
        name: app_workload
        required: true
        schema:
          title: App Workload
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/K8sMultiAgentSystemCRDViewModel"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get K8S Mas By App Workload
      tags:
      - Kubernetes Resources
  /k8s/namespaces/{namespace}/cache/store-token:
    post:
      operationId: cache_store_token
      parameters:
      - explode: false
        in: path
        name: namespace
        required: true
        schema:
          title: Namespace
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CacheTokenStoreRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Store Token
      tags:
      - Kubernetes Resources
  /k8s/namespaces/{namespace}/cache/load-token:
    post:
      operationId: cache_load_token
      parameters:
      - explode: false
        in: path
        name: namespace
        required: true
        schema:
          title: Namespace
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CacheTokenLoadRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Load Token
      tags:
      - Kubernetes Resources
...
  /k8s/trace/llm/call_end:
    post:
      description: Record the end of an LLM call for the authenticated agent.
      operationId: k8s_trace_llm_call_end
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LLMCallEndedKubernetesRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LLMCallEndedEvent"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: K8S Trace Llm Call End
      tags:
      - Kubernetes Resources
components:
  schemas:
    ActorClaim:
      description: |-
        Pydantic model for the JWT 'act' (actor) claim.

        Represents an actor in a delegation chain. Can be nested to represent
        a chain of delegation where the outermost act claim represents the
        current actor and nested act claims represent prior actors.

        As per RFC 8693, for access control decisions, only the top-level
        claims and the current actor (outermost act claim) should be considered.
      properties:
        sub:
          title: Sub
          type: string
        act:
          allOf:
          - $ref: "#/components/schemas/ActorClaim"
          nullable: true
      required:
      - sub
      title: ActorClaim
...
    AppMetadataResponse:
      description: Pydantic model for app metadata response.
      example:
        grant_types:
        - grant_types
        - grant_types
        jwks_uri: jwks_uri
        client_name: client_name
        client_id: client_id
        token_endpoint_auth_method: token_endpoint_auth_method
        response_types:
        - response_types
        - response_types
      properties:
        client_id:
          title: Client Id
          type: string
        client_name:
          title: Client Name
          type: string
        grant_types:
          items:
            type: string
          title: Grant Types
          type: array
        response_types:
          items:
            type: string
          title: Response Types
          type: array
        token_endpoint_auth_method:
          title: Token Endpoint Auth Method
          type: string
        jwks_uri:
          title: Jwks Uri
          type: string
      required:
      - client_id
      - client_name
      - grant_types
      - jwks_uri
      - response_types
      - token_endpoint_auth_method
      title: AppMetadataResponse
...
    AppType:
      description: Enumeration of app types.
      enum:
      - agent
      - client
      - mcp_server
      title: AppType
      type: string
...
    Body_introspect:
      properties:
        token:
          title: Token
          type: string
        tools:
          items:
            type: string
          nullable: true
          type: array
      required:
      - token
      title: Body_introspect
    CacheTokenLoadRequest:
      example:
        app_host: app_host
        trace_id: trace_id
        app_type: agent
        tool: tool
      properties:
        trace_id:
          title: Trace Id
          type: string
        app_host:
          title: App Host
          type: string
        app_type:
          $ref: "#/components/schemas/AppType"
        tool:
          nullable: true
          type: string
      required:
      - app_host
      - app_type
      - trace_id
      title: CacheTokenLoadRequest
    CacheTokenStoreRequest:
      example:
        app_host: app_host
        access_token: access_token
        trace_id: trace_id
        app_type: agent
        tool: tool
      properties:
        trace_id:
          title: Trace Id
          type: string
        app_host:
          title: App Host
          type: string
        app_type:
          $ref: "#/components/schemas/AppType"
        access_token:
          title: Access Token
          type: string
        tool:
          nullable: true
          type: string
      required:
      - access_token
      - app_host
      - app_type
      - trace_id
      title: CacheTokenStoreRequest
...
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - Location_inner
          - Location_inner
          input: ""
          ctx: "{}"
          type: type
        - msg: msg
          loc:
          - Location_inner
          - Location_inner
          input: ""
          ctx: "{}"
          type: type
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          title: Detail
          type: array
      title: HTTPValidationError
...
    K8sAppSpecViewModel:
      description: View model for application specification data.
      example:
        kubernetes_workload_name: kubernetes_workload_name
        name: name
        prompt_field_json_path: prompt_field_json_path
        mas_crd_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        type: agent
        app_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        url_host: url_host
        url_scheme: url_scheme
      properties:
        id:
          format: uuid
          nullable: true
          type: string
        name:
          title: Name
          type: string
        type:
          $ref: "#/components/schemas/AppType"
        url_host:
          title: Url Host
          type: string
        url_scheme:
          title: Url Scheme
          type: string
        prompt_field_json_path:
          nullable: true
          type: string
        kubernetes_workload_name:
          nullable: true
          type: string
        mas_crd_id:
          format: uuid
          nullable: true
          type: string
        app_id:
          format: uuid
          nullable: true
          type: string
      required:
      - name
      - type
      - url_host
      - url_scheme
      title: K8sAppSpecViewModel
...
    K8sMultiAgentSystemCRDViewModel:
      description: View model for a complete MultiAgentSystem CRD.
      example:
        enabled_tool_checks:
        - enabled_tool_checks
        - enabled_tool_checks
        llm_host: llm_host
        mas_metadata: ""
        kind: kind
        mas_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        namespace: namespace
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        app_specs:
        - kubernetes_workload_name: kubernetes_workload_name
          name: name
          prompt_field_json_path: prompt_field_json_path
          mas_crd_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          type: agent
          app_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          url_host: url_host
          url_scheme: url_scheme
        - kubernetes_workload_name: kubernetes_workload_name
          name: name
          prompt_field_json_path: prompt_field_json_path
          mas_crd_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          type: agent
          app_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          url_host: url_host
          url_scheme: url_scheme
        api_version: api_version
      properties:
        id:
          format: uuid
          nullable: true
          type: string
        api_version:
          title: Api Version
          type: string
        kind:
          title: Kind
          type: string
        namespace:
          title: Namespace
          type: string
        mas_metadata:
          allOf:
          - $ref: "#/components/schemas/K8sMultiAgentSystemMetadataViewModel"
          nullable: true
        name:
          title: Name
          type: string
        enabled_tool_checks:
          items:
            type: string
          nullable: true
          type: array
        llm_host:
          nullable: true
          type: string
        app_specs:
          default: []
          items:
            $ref: "#/components/schemas/K8sAppSpecViewModel"
          title: App Specs
          type: array
        mas_id:
          format: uuid
          nullable: true
          type: string
      required:
      - api_version
      - enabled_tool_checks
      - kind
      - llm_host
      - name
      - namespace
      title: K8sMultiAgentSystemCRDViewModel
    K8sMultiAgentSystemMetadataViewModel:
      description: View model for MultiAgentSystem metadata.
      properties:
        id:
          format: uuid
          nullable: true
          type: string
        name:
          title: Name
          type: string
        uid:
          nullable: true
          type: string
        resource_version:
          nullable: true
          type: string
        generation:
          nullable: true
          type: integer
        mas_crd_id:
          format: uuid
          nullable: true
          type: string
      required:
      - name
      title: K8sMultiAgentSystemMetadataViewModel
    LLMCallEndedEvent:
      description: Event emitted when an LLM call completes (recorded by the agent).
      example:
        user_input_id: user_input_id
        response: response
        mas_id: mas_id
        created_at: 2000-01-23T04:56:07.000+00:00
        id: id
        app_id: app_id
        tools: tools
        call_id: call_id
        token: token
      properties:
        id:
          title: Id
          type: string
        user_input_id:
          title: User Input Id
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        call_id:
          title: Call Id
          type: string
        token:
          title: Token
          type: string
        app_id:
          title: App Id
          type: string
        mas_id:
          nullable: true
          type: string
        response:
          title: Response
          type: string
        tools:
          nullable: true
          type: string
      required:
      - app_id
      - call_id
      - response
      - token
      - tools
      - user_input_id
      title: LLMCallEndedEvent
    LLMCallEndedKubernetesRequest:
      description: Request body for recording the end of an LLM call.
      example:
        response: response
        call_id: call_id
      properties:
        call_id:
          title: Call Id
          type: string
        response:
          title: Response
          type: string
      required:
      - call_id
      - response
      title: LLMCallEndedKubernetesRequest
    LLMCallEndedRequest:
      description: Request body for recording the end of an LLM call.
      example:
        response: response
        tools: tools
        call_id: call_id
      properties:
        call_id:
          title: Call Id
          type: string
        response:
          title: Response
          type: string
        tools:
          nullable: true
          type: string
      required:
      - call_id
      - response
      title: LLMCallEndedRequest
    LLMCallStartedEvent:
      description: Event emitted when an LLM call begins (recorded by the agent).
      example:
        user_input_id: user_input_id
        mas_id: mas_id
        created_at: 2000-01-23T04:56:07.000+00:00
        id: id
        app_id: app_id
        prompt: prompt
        tools: tools
        call_id: call_id
        token: token
      properties:
        id:
          title: Id
          type: string
        user_input_id:
          title: User Input Id
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        call_id:
          title: Call Id
          type: string
        token:
          title: Token
          type: string
        app_id:
          title: App Id
          type: string
        mas_id:
          nullable: true
          type: string
        prompt:
          title: Prompt
          type: string
        tools:
          nullable: true
          type: string
      required:
      - app_id
      - call_id
      - prompt
      - token
      - tools
      - user_input_id
      title: LLMCallStartedEvent
    LLMCallStartedRequest:
      description: Request body for recording the start of an LLM call.
      example:
        prompt: prompt
        tools: tools
        call_id: call_id
      properties:
        call_id:
          title: Call Id
          type: string
        prompt:
          title: Prompt
          type: string
        tools:
          nullable: true
          type: string
      required:
      - call_id
      - prompt
      title: LLMCallStartedRequest
...
    TokenExchangeRequest:
      description: A model representing a token exchange request.
      properties:
        client_id:
          title: Client Id
          type: string
        client_secret:
          title: Client Secret
          type: string
        subject_token:
          title: Subject Token
          type: string
        subject_token_type:
          title: Subject Token Type
          type: string
        scope:
          nullable: true
          type: string
        mcp_server_url:
          nullable: true
          type: string
        tools:
          items:
            type: string
          nullable: true
          type: array
      required:
      - client_id
      - client_secret
      - subject_token
      - subject_token_type
      title: TokenExchangeRequest
    TokenIntrospectResponse:
      description: Pydantic model for the token introspection response.
      example:
        sub: sub
        other:
          key: ""
        act: ""
        user_input_id: user_input_id
        scope: scope
        mas_id: mas_id
        active: true
        exp: 0
        app_id: app_id
        tools:
        - tools
        - tools
        client_id: client_id
      properties:
        client_id:
          nullable: true
          type: string
        scope:
          nullable: true
          type: string
        sub:
          nullable: true
          type: string
        act:
          allOf:
          - $ref: "#/components/schemas/ActorClaim"
          nullable: true
        other:
          additionalProperties: true
          nullable: true
          type: object
        exp:
          nullable: true
          type: integer
        user_input_id:
          nullable: true
          type: string
        app_id:
          nullable: true
          type: string
        mas_id:
          nullable: true
          type: string
        tools:
          items:
            type: string
          nullable: true
          type: array
        active:
          title: Active
          type: boolean
      required:
      - active
      title: TokenIntrospectResponse
    TokenRequest:
      description: A model representing a token generation request.
      properties:
        client_id:
          title: Client Id
          type: string
        client_secret:
          title: Client Secret
          type: string
        user_input:
          nullable: true
          type: string
        user_input_id:
          nullable: true
          type: string
      required:
      - client_id
      - client_secret
      title: TokenRequest
    TokenResponse:
      description: Pydantic model for the token response.
      example:
        access_token: access_token
      properties:
        access_token:
          title: Access Token
          type: string
      required:
      - access_token
      title: TokenResponse
...
    ValidationError:
      example:
        msg: msg
        loc:
        - Location_inner
        - Location_inner
        input: ""
        ctx: "{}"
        type: type
      properties:
        loc:
          items:
            $ref: "#/components/schemas/Location_inner"
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
        input:
          title: Input
        ctx:
          title: Context
          type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    Location_inner:
      anyOf:
      - type: string
      - type: integer
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http